// backend/@types/model.d.ts
import { Document } from "mongoose";

//? Type safety of the the user
export interface IUser extends Document {
  name: string;
  email: string;
  password: string;
  avatar: {
    public_id: string;
    url: string;
  };
  role: string;
  isVerified: boolean;
  courses: Array<{ courseId: string }>;
  comparePassword: (pass: string) => Promise<boolean>;
  SignAccessToken: () => string;
  SignRefreshToken: () => string;
}